Added some more functionality and tests
Completed the base functionality for the first frontend test (so that is next up).
The ability to edit threads and posts is now implemented and basic test coverage
for that is done.  The testting can probably be vastly improved, but I lack a lot
experience to know ahead of time what I am lacking.  Hopefully in the frontend stuff
I will be able to find holes and improve this.