Cleaned up responses some
Cleaned up the responses some, the threads now come back as just an array and not a
single object with a key to the array, the creation calls now return an id reference
to the created object.  The tests were updated to reflect these changes.  Also,
removed all of the trailing slashes in the URLs as flask was redirecting outside of
the subdirectory I was serving it from, redirects are evil...