Some random fixes from the browser runthrough
Fixed sessions complaining in the browser, not sure why it didn't show up in the
unittesting, but something was weird with the session key getting set in the cookies
but mongo returning a None.  Then switched to using a global db connection versus
generating a new one per module/file, should be nicer for mongo.  Also made the
response structures to be a bit more verbose.

I will be cleaning up the initial backbone sample at some point when I finish adding
some styles and make it usable (right now it looks like ass and is pretty confusing
to use).