Added settings options for date formatting
Added the setting for default session settings, these are added to a new Session
and can be overwritten with the /settings call.  Then applied this to how dates
come out in JSON by creating a new JSONEncoder that outputs the various formats,
the ones right now are ISO, seconds since epoch, and then strftime for anything
else.  This has plenty to be improved upon.