Added forums functionality
Added the forum (sub-forum) functionality, now threads can be placed in the root or
in a sub-forum, as well as retrieve threads from the root or subforum.  This raised
issues with how permissions were checked and required some new routes, so added
those too.  Fixed a few small things as well.