Major reorganization and rework of stuff
Modelled after the system used for another project, this moves everything up one
level from the /src directory, then most of the imports within the project are
relative and the big files have been split up.  All of the routes are via an initial
discovery packet or per object returned on a request.  Added the decorator stuff
from the other project as well.  Documentation needs to be updated, some more
cleaning should also be done.