Added the first try at a JS library
I really don't like this, the bind/trigger style is stupid.  I will probably try to
write a more OO version to use with the main frontend.  If I like that, I will try
to migrate it into use as the library.  Lots of corners cut and bad coverage of the
various return codes available, also the various wrappers are incredibly lax and
generic.